requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
beautifulsoup4==4.14.3
blinker==1.9.0
blis==1.3.3
cachetools==7.0.5
catalogue==2.0.10
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
cloudpathlib==0.23.0
confection==1.3.3
cymem==2.0.13
dataclasses-json==0.6.7
distro==1.9.0
fastapi==0.135.2
feedparser==6.0.12
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.46
graphviz==0.21
groq==0.37.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kuzu==0.11.3
langchain==1.2.13
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.22
langchain-groq==1.1.2
langchain-openai==1.1.12
langchain-text-splitters==1.1.1
langgraph==1.1.3
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.22
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
multidict==6.7.1
murmurhash==1.0.15
mypy_extensions==1.1.0
narwhals==2.18.1
networkx==3.6.1
numpy==2.4.3
openai==2.30.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pandas==2.3.3
pillow==12.1.1
plotly==6.6.0
preshed==3.0.13
propcache==0.4.1
protobuf==6.33.6
pyarrow==23.0.1
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
rdflib==7.6.0
referencing==0.37.0
regex==2026.2.28
requests==2.33.0
requests-toolbelt==1.0.0
rich==14.3.3
rpds-py==0.30.0
setuptools==82.0.1
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smart_open==7.5.1
smmap==5.0.3
sniffio==1.3.1
soupsieve==2.8.3
spacy==3.8.13
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.48
srsly==2.5.3
starlette==1.0.0
streamlit==1.55.0
streamlit-agraph==0.0.45
streamlit-option-menu==0.4.0
tenacity==9.1.4
thinc==8.3.13
tiktoken==0.12.0
toml==0.10.2
tornado==6.5.5
tqdm==4.67.3
typer==0.24.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.42.0
uvloop==0.22.1
wasabi==1.1.3
watchfiles==1.1.1
weasel==1.0.0
websockets==16.0
wrapt==2.1.2
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0
